pyproject.toml
[project]
name = "uorf-finder"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.86",
]
